# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgriMarket South Sudan API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, farmer, retailer]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: farmerId
          in: query
          schema:
            type: integer
        - name: available
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: currency
          in: query
          schema:
            type: string
            enum: [SSP, USD]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (farmer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (farmer/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/toggle-availability:
    patch:
      operationId: toggleProductAvailability
      tags: [products]
      summary: Toggle product availability (farmer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /trucks:
    get:
      operationId: listTrucks
      tags: [trucks]
      summary: List trucks with GPS location
      responses:
        "200":
          description: Trucks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Truck"
    post:
      operationId: createTruck
      tags: [trucks]
      summary: Register a truck (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TruckInput"
      responses:
        "201":
          description: Created truck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
  /trucks/{id}:
    get:
      operationId: getTruck
      tags: [trucks]
      summary: Get truck by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Truck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
    patch:
      operationId: updateTruck
      tags: [trucks]
      summary: Update truck location/status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TruckUpdate"
      responses:
        "200":
          description: Updated truck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, phone, role, language, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, farmer, retailer]
        farmName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        locationLat:
          type: ["number", "null"]
        locationLng:
          type: ["number", "null"]
        language:
          type: string
          enum: [en, ar]
        currency:
          type: string
          enum: [SSP, USD]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        farmName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        locationLat:
          type: ["number", "null"]
        locationLng:
          type: ["number", "null"]
        language:
          type: string
          enum: [en, ar]
        currency:
          type: string
          enum: [SSP, USD]
        isActive:
          type: boolean
    UserListResponse:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
...
    Product:
      type: object
      required: [id, name, nameAr, categoryId, farmerId, quantity, unit, priceSSP, priceUSD, available, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        farmerId:
          type: integer
        farmerName:
          type: ["string", "null"]
        farmName:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        priceSSP:
          type: number
        priceUSD:
          type: number
        available:
          type: boolean
        harvestDate:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        qualityGrade:
          type: string
          enum: [A, B, C]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, nameAr, categoryId, quantity, unit, priceSSP, priceUSD]
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        quantity:
          type: number
        unit:
          type: string
        priceSSP:
          type: number
        priceUSD:
          type: number
        harvestDate:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        qualityGrade:
          type: string
          enum: [A, B, C]
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        quantity:
          type: number
        unit:
          type: string
        priceSSP:
          type: number
        priceUSD:
          type: number
        available:
          type: boolean
        harvestDate:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        qualityGrade:
          type: string
          enum: [A, B, C]
    ProductListResponse:
      type: object
      required: [products, total]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
...
    Truck:
      type: object
      required: [id, plateNumber, driverName, driverPhone, status, createdAt]
      properties:
        id:
          type: integer
        plateNumber:
          type: string
        driverName:
          type: string
        driverPhone:
          type: string
        status:
          type: string
          enum: [available, in_transit, maintenance]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        currentOrderId:
          type: ["integer", "null"]
        createdAt:
          type: string
    TruckInput:
      type: object
      required: [plateNumber, driverName, driverPhone]
      properties:
        plateNumber:
          type: string
        driverName:
          type: string
        driverPhone:
          type: string
    TruckUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [available, in_transit, maintenance]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        driverName:
          type: string
        driverPhone:
          type: string
...