# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriMarket South Sudan API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductQualityGrade } from "./productQualityGrade";

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  /** @nullable */
  description?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  farmerId: number;
  /** @nullable */
  farmerName?: string | null;
  /** @nullable */
  farmName?: string | null;
  quantity: number;
  unit: string;
  priceSSP: number;
  priceUSD: number;
  available: boolean;
  /** @nullable */
  harvestDate?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  qualityGrade?: ProductQualityGrade;
  createdAt: string;
}