# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriMarket South Sudan API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputQualityGrade } from "./productInputQualityGrade";

export interface ProductInput {
  name: string;
  nameAr: string;
  /** @nullable */
  description?: string | null;
  categoryId: number;
  quantity: number;
  unit: string;
  priceSSP: number;
  priceUSD: number;
  /** @nullable */
  harvestDate?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  qualityGrade?: ProductInputQualityGrade;
}