# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriMarket South Sudan API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateCurrency } from "./userUpdateCurrency";
import type { UserUpdateLanguage } from "./userUpdateLanguage";

export interface UserUpdate {
  name?: string;
  /** @nullable */
  farmName?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  locationLat?: number | null;
  /** @nullable */
  locationLng?: number | null;
  language?: UserUpdateLanguage;
  currency?: UserUpdateCurrency;
  isActive?: boolean;
}